requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
